# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/history:
    get:
      operationId: getUserHistory
      tags: [user]
      summary: Get the authenticated user's private history
      responses:
        "200":
          description: Search, viewed-item, and barcode-scan history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserHistory"
        "401": { description: Authentication required }
        "403": { description: User is not approved }
        "500": { description: History could not be loaded }
    patch:
      operationId: updateUserHistory
      tags: [user]
      summary: Update one or more of the authenticated user's history collections
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserHistoryPatch"
      responses:
        "200":
          description: Updated history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserHistory"
        "400": { description: Invalid history update }
        "401": { description: Authentication required }
        "403": { description: User is not approved }
        "500": { description: History could not be saved }
  /user/me:
    delete:
      operationId: deleteUserMe
      tags: [auth]
      summary: Delete the authenticated user's own account
      description: >
        Deletes the caller's local DB row and Clerk identity.
        Returns 204 on success. If the DB delete succeeds but Clerk deletion
        fails, a 502 is returned with an error body so the caller knows the
        Clerk identity is still live.
      responses:
        "204":
          description: Account deleted successfully
        "400":
          description: Account cannot be deleted (e.g. bootstrap admin)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        "401":
          description: Authentication required
        "404":
          description: User record not found
        "502":
          description: DB deleted but Clerk deletion failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
...
components:
  schemas:
    UserHistory:
      type: object
      additionalProperties: false
      properties:
        queryHistory:
          type: array
          maxItems: 10
          items: { type: string, minLength: 1 }
        viewedHistory:
          type: array
          maxItems: 10
          items: { $ref: "#/components/schemas/UserHistoryViewedEntry" }
        scanHistory:
          type: array
          maxItems: 50
          items: { $ref: "#/components/schemas/UserHistoryScanEntry" }
      required: [queryHistory, viewedHistory, scanHistory]
    UserHistoryPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        queryHistory:
          type: array
          maxItems: 10
          items: { type: string, minLength: 1 }
        viewedHistory:
          type: array
          maxItems: 10
          items: { $ref: "#/components/schemas/UserHistoryViewedEntry" }
        scanHistory:
          type: array
          maxItems: 50
          items: { $ref: "#/components/schemas/UserHistoryScanEntry" }
    UserHistoryViewedEntry:
      type: object
      additionalProperties: false
      properties:
        id: { type: integer, minimum: 1 }
        catalog: { type: string }
        name: { type: string }
        vendor: { type: string }
        timestamp: { type: string, format: date-time }
      required: [id, catalog, name, vendor, timestamp]
    UserHistoryScanEntry:
      type: object
      additionalProperties: false
      properties:
        barcode: { type: string, minLength: 1 }
        found: { type: boolean }
        itemId: { type: integer, minimum: 1 }
        catalog: { type: string }
        vendor: { type: string }
        timestamp: { type: string, format: date-time }
        adminAction: { type: string, enum: [linked, created] }
      required: [barcode, found, timestamp]
...